+++ app/components/map/legend/sortable/list/item/component.tsx
