+++ app/components/map/legend/sortable/list/item/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
